[TensorExt] Add Interface method for resolve sliced accesses to a sparse tensor.
To resolve sliced access to a sparse tensor/memref (through
memref.subview/tensor.extract_slice/vector.transfer_read), a new
interface method, `resolveRange` is added to
`SparseOpInterface`. Given a Range this method can be used to resolve
the Range that represents the slice of the result into a range that
represents slice of the source.

Currently this method is only implemented for vector.transfer_read
operations.

Signed-off-by: MaheshRavishankar <mahesh.ravishankar@gmail.com>