[TensorExt] Add Interface methods for lowering to loops and estimating bounds.
This PR adds a couple of methods to the `SparseOpInterface` to allow for

- `lowerLoopRange` : Given a range of the rectangular domain of the
result of sparse operations accessed, generate a loop nest that
iterates over the actual (sparse) data in that range.

- `getEstimatedLoopRange` : Given a range of the rectangular domain of
the result of sparse operations accessed, generate an estimate of the
sparse range without introducing dependencies between the different
dimensions of the sparse data range.

The PR also adds a test pass that allows to verify the implementations
of these methods for the `cast_to_ragged_shape` operation.

Signed-off-by: MaheshRavishankar <mahesh.ravishankar@gmail.com>